Separated Discord/Capture polling, proper HTTP request cancellation on StopPolling
