+++ pkg/client/client.go
@@ -1,7 +1,6 @@
 package galactus_client
 
 import (
-	"bytes"
 	"encoding/json"
 	"errors"
 	"github.com/automuteus/galactus/pkg/discord_message"
@@ -10,9 +9,7 @@ import (
 	"github.com/automuteus/utils/pkg/capture"
 	"github.com/bwmarrin/discordgo"
 	"go.uber.org/zap"
-	"io/ioutil"
 	"net/http"
-	"time"
 )
 
 type GalactusClient struct {
@@ -60,111 +57,6 @@ func NewGalactusClient(address string, logger *zap.Logger) (*GalactusClient, err
 		return &gc, errors.New("galactus returned a non-200 status code; ensure it is reachable")
 	}
 	return &gc, nil
-}
-
-type PollingType int
-
-const (
-	DiscordPolling PollingType = iota
-	CapturePolling
-)
-
-func (galactus *GalactusClient) StartPolling(pollingType PollingType, connectCode string) error {
-	var channel chan struct{}
-	if pollingType == CapturePolling {
-		valid, err := validate.ValidConnectCode(connectCode)
-		if !valid {
-			return err
-		}
-		if _, ok := galactus.captureKillChannels[connectCode]; ok {
-			return errors.New("already polling for capture events for connect code " + connectCode)
-		}
-		galactus.captureKillChannels[connectCode] = make(chan struct{})
-		channel = galactus.captureKillChannels[connectCode]
-	} else {
-		if galactus.discordKillChannel != nil {
-			return errors.New("already polling for discord events")
-		}
-		galactus.discordKillChannel = make(chan struct{})
-		channel = galactus.discordKillChannel
-	}
-
-	connected := false
-	requestCancelChannel := make(chan struct{})
-
-	go func() {
-		for {
-			select {
-			case <-channel:
-				requestCancelChannel <- struct{}{}
-				return
-
-			default:
-				var url string
-				switch pollingType {
-				case DiscordPolling:
-					url = endpoint.FormGalactusURL(galactus.Address, endpoint.DiscordRoute, endpoint.DiscordJobRequest)
-				case CapturePolling:
-					url = endpoint.FormGalactusURL(galactus.Address, endpoint.CaptureRoute, endpoint.GetCaptureEventPartial, connectCode)
-				}
-				req, err := http.NewRequest("POST", url, bytes.NewBufferString(""))
-				if err != nil {
-					galactus.logger.Error("invalid url provided to galactus client",
-						zap.String("url", url))
-					break
-				}
-				req.Cancel = requestCancelChannel
-
-				response, err := http.DefaultClient.Do(req)
-				if err != nil {
-					connected = false
-					galactus.logger.Error("could not reach galactus",
-						zap.Error(err),
-						zap.String("url", url))
-					galactus.logger.Info("waiting 1 second before retrying")
-					time.Sleep(time.Second * 1)
-				} else {
-					if !connected {
-						galactus.logger.Info("successful connection to galactus")
-						connected = true
-					}
-					body, err := ioutil.ReadAll(response.Body)
-					if err != nil {
-						galactus.logger.Error("error reading http response from galactus",
-							zap.Error(err),
-							zap.String("url", url),
-							zap.ByteString("message", body))
-					} else if response.StatusCode == http.StatusOK {
-						switch pollingType {
-						case DiscordPolling:
-							var msg discord_message.DiscordMessage
-							err := json.Unmarshal(body, &msg)
-							if err != nil {
-								galactus.logger.Error("error unmarshalling discord message from galactus",
-									zap.Error(err),
-									zap.ByteString("message", body))
-							} else {
-								galactus.dispatchDiscordMessage(msg)
-							}
-						case CapturePolling:
-							var msg capture.Event
-							err := json.Unmarshal(body, &msg)
-							if err != nil {
-								galactus.logger.Error("error unmarshalling capture message from galactus",
-									zap.Error(err),
-									zap.ByteString("message", body))
-							} else {
-								galactus.dispatchCaptureMessage(connectCode, msg)
-							}
-						}
-
-					}
-					response.Body.Close()
-				}
-			}
-		}
-	}()
-	return nil
 }
 
 func (galactus *GalactusClient) dispatchDiscordMessage(msg discord_message.DiscordMessage) {
@@ -240,27 +132,6 @@ func (galactus *GalactusClient) dispatchCaptureMessage(connectCode string, msg c
 	}
 }
 
-func (galactus *GalactusClient) StopCapturePolling(connectCode string) {
-	if galactus.captureKillChannels[connectCode] != nil {
-		galactus.captureKillChannels[connectCode] <- struct{}{}
-	}
-	delete(galactus.captureKillChannels, connectCode)
-}
-
-func (galactus *GalactusClient) StopDiscordPolling() {
-	if galactus.discordKillChannel != nil {
-		galactus.discordKillChannel <- struct{}{}
-	}
-	galactus.discordKillChannel = nil
-}
-
-func (galactus *GalactusClient) StopAllPolling() {
-	galactus.StopDiscordPolling()
-	for i := range galactus.captureKillChannels {
-		galactus.StopCapturePolling(i)
-	}
-}
-
 func (galactus *GalactusClient) RegisterDiscordHandler(msgType discord_message.DiscordMessageType, f interface{}) bool {
 	registered := false
 	switch msgType {
@@ -291,6 +162,9 @@ func (galactus *GalactusClient) RegisterDiscordHandler(msgType discord_message.D
 }
 
 func (galactus *GalactusClient) RegisterCaptureHandler(connectCode string, f interface{}) bool {
+	if len(connectCode) != validate.ConnectCodeLength {
+		return false
+	}
 	if handlers, ok := galactus.genericCaptureHandlers[connectCode]; ok {
 		handlers = append(handlers, f.(func(msg capture.Event)))
 		galactus.genericCaptureHandlers[connectCode] = handlers
